+++ types/node/crypto.d.ts
@@ -261,7 +261,7 @@ declare module "crypto" {
     }
 
     function createPrivateKey(key: PrivateKeyInput | string | Buffer): KeyObject;
-    function createPublicKey(key: PublicKeyInput | string | Buffer): KeyObject;
+    function createPublicKey(key: PublicKeyInput | string | Buffer | KeyObject): KeyObject;
     function createSecretKey(key: Buffer): KeyObject;
 
     function createSign(algorithm: string, options?: stream.WritableOptions): Signer;
