+++ types/node/events.d.ts
@@ -5,6 +5,7 @@ declare module "events" {
          class EventEmitter extends internal {
             /** @deprecated since v4.0.0 */
             static listenerCount(emitter: EventEmitter, event: string | symbol): number;
+            static once(emitter: EventEmitter, event: string | symbol): Promise<any>;
             static defaultMaxListeners: number;
 
             addListener(event: string | symbol, listener: (...args: any[]) => void): this;
