+++ types/node/tsconfig.json
@@ -50,6 +50,7 @@
         "test/cluster.ts",
         "test/crypto.ts",
         "test/dgram.ts",
+        "test/events.ts",
         "test/global.ts",
         "test/http.ts",
         "test/http2.ts",
@@ -65,6 +66,7 @@
         "test/tls.ts",
         "test/tty.ts",
         "test/util.ts",
+        "test/v8.ts",
         "test/worker_threads.ts",
         "test/zlib.ts"
     ],
